fix(checkout): moved redirect to the correct page
