+++ src/pages/checkout/index.jsx
@@ -1,4 +1,7 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { connect } from 'react-redux';
+import { navigate } from 'gatsby';
+import { isPurchaseCreated } from 'data/checkout/selectors';
 import { CheckoutMenu } from 'components/Menu';
 import SectionLayout, { SectionTitle } from 'layouts/section';
 import { TicketSelector } from 'components/TicketSelector';
@@ -22,18 +25,30 @@ const CHECKOUT_MENU = [
   }
 ];
 
-const CheckoutPage = () => (
-  <SectionLayout
-    title="Entradas"
-    className={styles.section}
-    menuComponent={() => <CheckoutMenu items={CHECKOUT_MENU} short />}
-  >
-    <SectionTitle>Entradas</SectionTitle>
-    <div className={styles.tickets}>
-      <TicketSelector />
-      <TicketForm />
-    </div>
-  </SectionLayout>
-);
+const CheckoutPage = ({ alreadyInitiated }) => {
+  useEffect(() => {
+    if (alreadyInitiated) {
+      navigate('/checkout/review/');
+    }
+  }, [alreadyInitiated]);
 
-export default CheckoutPage;
+  return (
+    <SectionLayout
+      title="Entradas"
+      className={styles.section}
+      menuComponent={() => <CheckoutMenu items={CHECKOUT_MENU} short />}
+    >
+      <SectionTitle>Entradas</SectionTitle>
+      <div className={styles.tickets}>
+        <TicketSelector />
+        <TicketForm />
+      </div>
+    </SectionLayout>
+  );
+};
+
+const mapStateToProps = state => ({
+  alreadyInitiated: isPurchaseCreated(state)
+});
+
+export default connect(mapStateToProps)(CheckoutPage);
