+++ src/pages/checkout/review.jsx
@@ -1,7 +1,4 @@
-import React, { useEffect } from 'react';
-import { connect } from 'react-redux';
-import { navigate } from 'gatsby';
-import { isPurchaseCreated } from 'data/checkout/selectors';
+import React from 'react';
 import { CheckoutMenu } from 'components/Menu';
 import SectionLayout, { SectionTitle } from 'layouts/section';
 import { PaymentReviewCard } from 'components/PaymentReviewCard';
@@ -24,31 +21,19 @@ const CHECKOUT_MENU = [
   }
 ];
 
-const CheckoutReviewPage = ({ alreadyInitiated }) => {
-  useEffect(() => {
-    if (alreadyInitiated) {
-      navigate('/checkout/review/');
-    }
-  }, [alreadyInitiated]);
+const CheckoutReviewPage = ({ alreadyInitiated }) => (
+  <SectionLayout
+    title="Revisá tu compra"
+    className={styles.section}
+    menuComponent={() => <CheckoutMenu items={CHECKOUT_MENU} short hideOnMobile />}
+    mobileBackButtonRoute="/checkout/"
+    hideFooterOnMobile
+  >
+    <SectionTitle>Entradas</SectionTitle>
+    <div className={styles.reviewContainer}>
+      <PaymentReviewCard />
+    </div>
+  </SectionLayout>
+);
 
-  return (
-    <SectionLayout
-      title="Revisá tu compra"
-      className={styles.section}
-      menuComponent={() => <CheckoutMenu items={CHECKOUT_MENU} short hideOnMobile />}
-      mobileBackButtonRoute="/checkout/"
-      hideFooterOnMobile
-    >
-      <SectionTitle>Entradas</SectionTitle>
-      <div className={styles.reviewContainer}>
-        <PaymentReviewCard />
-      </div>
-    </SectionLayout>
-  );
-};
-
-const mapStateToProps = state => ({
-  alreadyInitiated: isPurchaseCreated(state)
-});
-
-export default connect(mapStateToProps)(CheckoutReviewPage);
+export default CheckoutReviewPage;
